Reorganize models into distinct files
This will be a better organization because we'll have one route per resource
