+++ src/rust/src/models/mod.rs
@@ -1,1 +1,3 @@
 pub mod card;
+pub mod deck;
+pub mod note;
